Start work on making rendering errors an explicit type of error.
It can be hard to understand why a particular template failed to render,
as the only types of errors returned by Tera are `Error::msg`s, which
don't have very much information in them. We add a `RenderErrorKind`
and `RenderError` variant which can encode information for why a
particular template failed to render. This also allows consuming
applications to more easily detect and respond to render errors in a
useful way (for example, displaying a better end-user message)